gtk2_ui: Create a GLib main-context for each thread
A main context represents a set of callbacks, sockets and other
event sources, and each main-context can be acquired (owned) by at
most one thread at a time. A main-context is normally owned by the
same thread for its entire lifetime, and that's how we use them here.

The "thread-default main context" is how GLib and GIO track the thread
in which the callback for an async operation should be called; we don't
actually use any GIO async operations at the moment, but it's a good
idea to set it correctly. This also gives us the ability to give better
diagnostics than just asserting is_owner(), because we can show both
the expected thread and the actual thread in the assertion message.