utils.py: Fix IndexError: list index out of range
re.findall() called with a regexp containing groups will return a list
of tuples. In our case: a list containing zero or exactly one tuple.

Since we are storing the result in a list anyway, simply add the two
lists and do not try to extract an element first. If re.findall()
returns an empty list, this does nothing and will no longer crash.