Detect runit as init
Reportbug has no code to detect runit as init; moreover both
sysvinit-core and runit-init ship a '/sbin/init' file, so there is the
chance that reportbug wrongly detect sysv as init when in fact runit is
in charge. This commit let reportbug detect when runit is init by
checking for the existence of '/run/runit.stopit' file. That file is
created during Stage-1 (early one-time boot tasks) and is used by runit
during Stage-3 (shutdown tasks).